console: Refactor pubsub edit view use update action instead of direct api call
